camera: move Camera_UpdateCutscene to TRX
This moves both games' Camera_UpdateCutscene functions to TRX as-is.
