+++ src/libtrx/game/camera/cinematic.c
@@ -2,6 +2,7 @@
 
 #include "game/camera.h"
 #include "game/game_buf.h"
+#include "game/lara.h"
 #include "game/rooms.h"
 #include "game/viewport.h"
 
@@ -93,5 +94,69 @@ void Camera_LoadCutsceneFrame(void)
 
     Viewport_AlterFOV(fov);
     Camera_UpdateMicPosition();
+#endif
+}
+
+void Camera_UpdateCutscene(void)
+{
+    const CINE_DATA *const cine_data = Camera_GetCineData();
+    if (cine_data->frame_count == 0) {
+        return;
+    }
+
+    const CINE_FRAME *const frame = Camera_GetCurrentCineFrame();
+#if TR_VERSION == 1
+    const int32_t c = Math_Cos(cine_data->position.rot.y);
+    const int32_t s = Math_Sin(cine_data->position.rot.y);
+    const XYZ_32 *const pos = &cine_data->position.pos;
+    g_Camera.target.x = pos->x + ((c * frame->tx + s * frame->tz) >> W2V_SHIFT);
+    g_Camera.target.y = pos->y + frame->ty;
+    g_Camera.target.z = pos->z + ((c * frame->tz - s * frame->tx) >> W2V_SHIFT);
+    g_Camera.pos.x = pos->x + ((s * frame->cz + c * frame->cx) >> W2V_SHIFT);
+    g_Camera.pos.y = pos->y + frame->cy;
+    g_Camera.pos.z = pos->z + ((c * frame->cz - s * frame->cx) >> W2V_SHIFT);
+    const int16_t room_num =
+        Room_GetIndexFromPos(g_Camera.pos.x, g_Camera.pos.y, g_Camera.pos.z);
+    if (room_num != NO_ROOM) {
+        g_Camera.pos.room_num = room_num;
+    }
+    g_Camera.roll = frame->roll;
+    g_Camera.shift = 0;
+
+    Viewport_SetFOV(frame->fov);
+#else
+    int32_t tx = frame->tx;
+    int32_t ty = frame->ty;
+    int32_t tz = frame->tz;
+    int32_t cx = frame->cx;
+    int32_t cy = frame->cy;
+    int32_t cz = frame->cz;
+    int32_t fov = frame->fov;
+    int32_t roll = frame->roll;
+    int32_t c = Math_Cos(g_Camera.target_angle);
+    int32_t s = Math_Sin(g_Camera.target_angle);
+
+    const ITEM *const lara_item = Lara_GetItem();
+    const XYZ_32 camera_target = {
+        .x = lara_item->pos.x + ((tx * c + tz * s) >> W2V_SHIFT),
+        .y = lara_item->pos.y + ty,
+        .z = lara_item->pos.z + ((tz * c - tx * s) >> W2V_SHIFT),
+    };
+    const XYZ_32 camera_pos = {
+        .x = lara_item->pos.x + ((cx * c + cz * s) >> W2V_SHIFT),
+        .y = lara_item->pos.y + cy,
+        .z = lara_item->pos.z + ((cz * c - cx * s) >> W2V_SHIFT),
+    };
+    const int16_t room_num =
+        Room_GetIndexFromPos(camera_pos.x, camera_pos.y, camera_pos.z);
+    if (room_num != NO_ROOM) {
+        g_Camera.pos.room_num = room_num;
+    }
+
+    g_Camera.pos.pos = camera_pos;
+    g_Camera.target.pos = camera_target;
+    g_Camera.roll = roll;
+    g_Camera.shift = 0;
+    Viewport_AlterFOV(fov);
 #endif
 }
